move ClearUserSessionUseCase to :core:domain
